Fixing bug (see #62 comments) and extracting ExecutionContext
