+++ index.html
@@ -109,8 +109,9 @@
 								<strong>Injected Web3</strong>
 								<p>Execution environment has been provided by Mist or similar provider.</p>
 							</label>
-							<label for="web3">
-								<input id="web3" type="radio" value="web3" name="executionContext">
+							<label for="_web3">
+								<!-- This element cannot be named web3 due to this HTML quirk: http://www.2ality.com/2012/08/ids-are-global.html. In order to detect when web3 is injected by Mist, MetaMask, etc, there cannot be any HTML elements with and id of web3. -->
+								<input id="_web3" type="radio" value="web3" name="executionContext">
 								<strong>Web3 Provider</strong>
 																<p>Execution environment connects to node at localhost (or via IPC if available), transactions will be sent to the network and can cause loss of money or worse!<br/>
 																<b>If this page is served via https and you access your node via http, it might not work. In this case, try cloning the repository and serving it via http.</b></p>
